mini_projects/dezsi.lajos/app/backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-google-genai
langchain-core
langgraph
qdrant-client
python-dotenv
httpx
